Removed obsolete methods and variables
