+++ src/main/java/com/bae/SpringConsumerApplication.java
@@ -25,109 +25,55 @@ public class SpringConsumerApplication {
 		SpringApplication.run(SpringConsumerApplication.class, args);
 	}
 
-//	@Bean
-//	public JmsListenerContainerFactory<?> myFactory(ConnectionFactory connectionFactory,
-//			DefaultJmsListenerContainerFactoryConfigurer configurer) {
-//		DefaultJmsListenerContainerFactory factory = new DefaultJmsListenerContainerFactory();
-//		configurer.configure(factory, connectionFactory);
-//		return factory;
-//	}
-//
-//	@Bean
-//	public MessageConverter jacksonJmsMessageConverter() {
-//		MappingJackson2MessageConverter converter = new MappingJackson2MessageConverter();
-//		converter.setTargetType(MessageType.TEXT);
-//		converter.setTypeIdPropertyName("account");
-//		return converter;
-//	}
-//
-//	@Bean
-//	public RestTemplate restTemplate(RestTemplateBuilder builder) {
-//		return builder.build();
-//	}
-
 	String BROKER_URL = "tcp://localhost:61616";
-
 	String BROKER_USERNAME = "admin";
-
 	String BROKER_PASSWORD = "admin";
 
 	@Bean
-
 	public ActiveMQConnectionFactory connectionFactory() {
-
 		ActiveMQConnectionFactory connectionFactory = new ActiveMQConnectionFactory();
 
 		connectionFactory.setBrokerURL(BROKER_URL);
-
 		connectionFactory.setPassword(BROKER_USERNAME);
-
 		connectionFactory.setUserName(BROKER_PASSWORD);
-
 		return connectionFactory;
-
 	}
 
 	@Bean
-
 	public JmsTemplate jmsTemplate() {
-
 		JmsTemplate template = new JmsTemplate();
-
 		template.setConnectionFactory(connectionFactory());
-
 		return template;
-
 	}
 
 	@Bean
-
 	public DefaultJmsListenerContainerFactory jmsListenerContainerFactory() {
-
 		DefaultJmsListenerContainerFactory factory = new DefaultJmsListenerContainerFactory();
-
 		factory.setConnectionFactory(connectionFactory());
-
 		factory.setConcurrency("1-1");
-
 		return factory;
-
 	}
 
 	@Bean
-
 	public JmsListenerContainerFactory<?> myFactory(ConnectionFactory connectionFactory,
-
 			DefaultJmsListenerContainerFactoryConfigurer configurer) {
 
 		DefaultJmsListenerContainerFactory factory = new DefaultJmsListenerContainerFactory();
-
 		configurer.configure(factory, connectionFactory);
-
 		return factory;
-
 	}
 
 	@Bean
-
 	public MessageConverter jacksonJmsMessageConverter() {
-
 		MappingJackson2MessageConverter converter = new MappingJackson2MessageConverter();
-
 		converter.setTargetType(MessageType.TEXT);
-
 		converter.setTypeIdPropertyName("_type");
-
 		return converter;
-
 	}
 
 	@Bean
-
 	public RestTemplate restTemplate(RestTemplateBuilder builder) {
-
 		return builder.build();
-
 	}
 
 }
